requirements.txt
pandas
requests
snowflake-connector-python
sqlalchemy
apache-airflow
psycopg2-binary